Change task handle type to intptr_t to match SWI

+++ SFToSpr/tests/Tests.c
@@ -180,7 +180,7 @@ typedef enum
 }
 DataTransferMethod;
 
-static int th;
+static intptr_t th;
 
 static void wipe(const char *path_name)
 {
